Revert "replacing old syntax with new"
This reverts commit 0413afc525a5bd9a682808df9c56748345a25a82.

+++ ExperPort/Modules/@dispatcher/RunningSection.m
@@ -19,7 +19,7 @@ switch action,
         % ------------------------------------------------------------------
         %%                       INIT
         % ------------------------------------------------------------------
-        SoloParamHandle(obj, 'my_gui_info', 'value', [x y double(gcf)]);
+        SoloParamHandle(obj, 'my_gui_info', 'value', [x y gcf]);
 
         SoloParamHandle(obj, 'Running', 'value', 0);
         PushbuttonParam(obj, 'RunButton', x, y, ...
